Add RegradeController
