+++ site/app/views/NavigationView.php
@@ -356,7 +356,7 @@ class NavigationView extends AbstractView {
             "(due " . $gradeable->getSubmissionDueDate()->format(self::DATE_FORMAT) . ")";
         $points_percent = NAN;
 
-        $href = $this->core->buildNewCourseUrl(['student', $gradeable->getId()]);
+        $href = $this->core->buildNewCourseUrl(['gradeable', $gradeable->getId()]);
         $progress = null;
         $disabled = false;
